Finished the basic tagging functionality + random refactoring

+++ misli_desktop/statisticsmodule.h
@@ -9,7 +9,7 @@ public:
     struct Statistic{
         QString name;
         QList<QDateTime> timestamps;
-        QList<float> values;
+        QList<double> values;
     };
 
     //Functions
